+++ config/config.exs
@@ -5,7 +5,5 @@
 import Config
 
 config :logger, level: :info
-# TODO: check whether we should replace it with something
-# config :fledex, Fledex.Animation.JobScheduler, debug_logging: false
 
 import_config "config_#{Mix.env()}.exs"